FIX: Wrong type in function param. UPDATE: .travis.yml

+++ src/DSPHelp.jl
@@ -1,11 +1,11 @@
 module DSPHelp
 export ev, od
 
-function ev(x::Array{Int64,2})
+function ev(x::Array{Float64,2})
   return (x + fliplr(x))/2
 end
 
-function od(x::Array{Int64,2})
+function od(x::Array{Float64,2})
   return (x - fliplr(x))/2
 end 
 # package code goes here
